refactor: fix the path of the folders and the name of the enum
