+++ src/products/entities/products.ts
@@ -1,5 +1,5 @@
 import { Document } from 'mongoose';
-import { StatusEnum } from '../../types/status-enum';
+import { StatusEnum } from '../types/status-enum';
 export class Product extends Document {
   name: string;
   price: number;
